aidevs3/w2/s02e04/requirements.txt
openai
openai-whisper
ollama
PyYAML
requests
easyocr
opencv-python
